+++ packages/breadboard/src/data/file-system/persistent-file.ts
@@ -14,34 +14,9 @@ import {
   FileSystemWriteResult,
   Outcome,
 } from "../types.js";
-import { err, ok } from "./utils.js";
+import { noStreams, ok, readFromStart } from "./utils.js";
 
 export { PersistentFile };
-
-// TODO: Move to common
-function readFromStart(
-  path: FileSystemPath,
-  data: LLMContent[] | undefined,
-  start: number
-): FileSystemReadResult {
-  if (!data) {
-    return err(`File at "${path}" is empty`);
-  }
-
-  if (start >= data.length) {
-    return err(`Length of file is lesser than start "${start}"`);
-  }
-  return {
-    data: data.slice(start),
-    last: data.length - 1,
-  };
-}
-
-function noStreams(done: boolean, receipt?: boolean): FileSystemWriteResult {
-  if (done || receipt) {
-    return err("Can't close the file that isn't a stream");
-  }
-}
 
 class PersistentFile implements FileSystemFile {
   constructor(
