+++ packages/breadboard/src/data/file-system/utils.ts
@@ -4,9 +4,15 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { Outcome } from "../types.js";
+import { LLMContent } from "@breadboard-ai/types";
+import {
+  FileSystemPath,
+  FileSystemReadResult,
+  FileSystemWriteResult,
+  Outcome,
+} from "../types.js";
 
-export { ok, err };
+export { ok, err, readFromStart, noStreams };
 
 function ok<T>(o: Outcome<T>): o is T {
   return !(o && typeof o === "object" && "$error" in o);
@@ -14,4 +20,28 @@ function ok<T>(o: Outcome<T>): o is T {
 
 function err($error: string) {
   return { $error };
+}
+
+function readFromStart(
+  path: FileSystemPath,
+  data: LLMContent[] | undefined,
+  start: number
+): FileSystemReadResult {
+  if (!data) {
+    return err(`File at "${path}" is empty`);
+  }
+
+  if (start >= data.length) {
+    return err(`Length of file is lesser than start "${start}"`);
+  }
+  return {
+    data: data.slice(start),
+    last: data.length - 1,
+  };
+}
+
+function noStreams(done: boolean, receipt?: boolean): FileSystemWriteResult {
+  if (done || receipt) {
+    return err("Can't close the file that isn't a stream");
+  }
 }
